package_manager: deb: update conffiles md5sum
Currently in an image created with debian package management, dpkg is
missing the md5sums in /var/lib/dpkg/status for files designated as
conffiles.  They are just listed as "newconffile", which is an
intermediate state when dpkg installs a package with an conf file.

Also, when the device first boots the script /usr/sbin/run-postinsts
does not cause /var/lib/dpkg/status to update those hashes either.

The problem here is that when using a package feed, all the conf files
are listed as "newconffile", which will indicate to dpkg to replace
them all on the first update.  dpkg will not follow the rules for
confold, confnew, confdef.

Original commit:
 https://patchwork.yoctoproject.org/project/oe-core/patch/20220726211328.2122681-1-rustyhowell@gmail.com/#4474

Signed-off-by: Jean-Marie Lemetayer <j.lemetayer@kerlink.fr>
Signed-off-by: Rusty Howell <rustyhowell@gmail.com>