package: add a way to mark packages as automatically installed
Using the PACKAGES_MARKED_AS_AUTO variable allows packages contained
in images to be marked as automatically installed.

This is useful for identifying packages installed by a future user, but
the package manager must be configured not to remove essential packages.

With Debian, these packages must be marked as "Essential" or be part of
the "APT::NeverAutoRemove" configuration.

https://www.debian.org/doc/debian-policy/ch-controlfields.html#essential
https://manpages.debian.org/testing/apt/apt.conf.5.en.html

This commit only implements the feature for Debian / APT.

Signed-off-by: Jean-Marie Lemetayer <j.lemetayer@kerlink.fr>
Signed-off-by: Guillaume Lecocq <g.lecocq@kerlink.fr>